Successfully implemented go to previous/next member, with some work still to be done.
