+++ HotCommands/HotCommands.csproj
@@ -59,7 +59,7 @@
     <Compile Include="Commands\FormatCode.cs" />
     <Compile Include="Commands\EditorExtensions.cs" />
     <Compile Include="Commands\ExpandSelection.cs" />
-    <Compile Include="Commands\MoveCursorToPreviousMember.cs" />
+    <Compile Include="Commands\MoveCursorToAdjacentMember.cs" />
     <Compile Include="Commands\MoveMemberDown.cs" />
     <Compile Include="Commands\MoveMemberUp.cs" />
     <Compile Include="Constants.cs" />
